ioctl: hide internal space_args function
The only reason this function exists is to reduce confusion when using
the same variable names etc. Getting space info is a two step process,
first we ask for the amount of space, and then we call again with a
larger memory buffer to hold the space info.

The _space_args must never be called from outside the ioctl module.